+++ src/client/ui/module/ui_init.c
@@ -131,6 +131,7 @@ void UI_Init(void)
     UI_LoadMenus("ui_mp/ingame.txt", qfalse, R_IMAGE_TRACK_UI);
     trap_Cvar_Set(UI_COMPAT_CONSOLE_BIND_CAPTURE_CVAR, "0");
     ui_compat_extend_graphics_menu();
+    ui_compat_extend_sound_menu();
     ui_compat_extend_console_binding_menu();
     ui_compat_extend_advanced_menu();
     ui_compat_brand_main_menu_version();